# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SkillScope – AI-Powered Developer Intelligence Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/me:
    get:
      operationId: getMe
      tags: [users]
      summary: Get current user profile
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Unauthorized
    put:
      operationId: updateMe
      tags: [users]
      summary: Update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserProfileUpdate"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
...
  /roadmap/goals/{id}/complete:
    patch:
      operationId: completeGoal
      tags: [roadmap]
      summary: Mark a goal as complete
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated goal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"

  # ─── Gamification ─────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    UserProfile:
      type: object
      required: [id, clerkId, email, role, createdAt]
      properties:
        id:
          type: integer
        clerkId:
          type: string
        email:
          type: string
        name:
          type: ["string", "null"]
        username:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        college:
          type: ["string", "null"]
        graduationYear:
          type: ["integer", "null"]
        role:
          type: string
          enum: [student, recruiter, admin]
        overallScore:
          type: ["integer", "null"]
        githubConnected:
          type: boolean
        xp:
          type: integer
          default: 0
        level:
          type: integer
          default: 1
        streak:
          type: integer
          default: 0
        createdAt:
          type: string
    UserProfileUpdate:
      type: object
      properties:
        name:
          type: string
        username:
          type: string
        bio:
          type: string
        location:
          type: string
        college:
          type: string
        graduationYear:
          type: integer

    # ─── Dashboard ──────────────────────────────────────────────────────────────
...
    Goal:
      type: object
      required: [id, title, type, isCompleted]
      properties:
        id:
          type: integer
        title:
          type: string
        type:
          type: string
          enum: [daily, weekly, monthly]
        isCompleted:
          type: boolean
        xpReward:
          type: integer

    # ─── Gamification ───────────────────────────────────────────────────────────
...